requirements.txt
Django==1.11.12
django-crispy-forms==1.7.2
django-filter==2.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
Markdown==2.6.11
pkg-resources==0.0.0
public==1.0.3
PyJWT==1.6.4
pytz==2018.5
